remove warnings in linux

+++ modules/CompVis/iACompBayesianBlocksData.cpp
@@ -17,7 +17,7 @@ void iACompBayesianBlocksData::calculateNumberOfObjectsInEachBin(
 	{//check for every dataset
 
 		auto currDataset = thisBinDataObjects->at(v);
-		auto currentNumberOfBins = currDataset->size();
+		auto currentNumberOfBins = static_cast<int>(currDataset->size());
 		
 		std::vector<double> bins;
